Fix Preempt Dir

+++ src/fairchem/core/components/train/train_runner.py
@@ -24,6 +24,7 @@ from fairchem.core.components.runner import Runner
 from fairchem.core.units.mlip_unit.mlip_unit import (
     convert_train_checkpoint_to_inference_checkpoint,
 )
+from fairchem.core._cli import PREEMPTION_STATE_DIR_NAME
 
 if TYPE_CHECKING:
     from torch.distributed.checkpoint.stateful import Stateful
@@ -84,7 +85,8 @@ class TrainCheckpointCallback(Callback):
                     self.checkpoint_dir
                 )
                 for dir, _ in checkpoint_dirs_by_time[: -self.max_saved_checkpoints]:
-                    shutil.rmtree(dir)
+                    if not dir.endswith(PREEMPTION_STATE_DIR_NAME):
+                        shutil.rmtree(dir)
 
     def on_train_end(self, state: State, unit: TTrainUnit) -> None:
         if self.checkpoint_every_n_steps is not None:
